Update service methods + remove methods

+++ src/main/java/backEnd1/pensionat/controllers/CustomerController.java
@@ -2,7 +2,10 @@ package backEnd1.pensionat.controllers;
 
 import backEnd1.pensionat.DTOs.SimpleCustomerDTO;
 import backEnd1.pensionat.Models.Customer;
+
+import backEnd1.pensionat.services.interfaces.BookingService;
 import backEnd1.pensionat.services.convert.CustomerConverter;
+
 import backEnd1.pensionat.services.interfaces.CustomerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -20,6 +23,8 @@ public class CustomerController {
 
     private final CustomerService customerService;
 
+    private final BookingService bookingService;
+
    /* @RequestMapping("/all")
     public List<Customer> getAllCustomers() {
         return customerService.getAllCustomers();
@@ -30,6 +35,11 @@ public class CustomerController {
         return "customerOrNot";
     }
 
+    @PostMapping("/addCustomerObject")
+    public Customer addCustomer(@RequestBody Customer c){
+        return customerService.addCustomer(c);
+    }
+
     @DeleteMapping("/{id}/remove")
     public String removeCustomerById(@PathVariable Long id) {
         return customerService.removeCustomerById(id);
@@ -38,22 +48,25 @@ public class CustomerController {
     //DeleteMapping ovan verkar ej fungera. Använder temp. Requestmapping.
     @RequestMapping("/{id}/removeHandler")
     public String removeCustomerByIdHandler(@PathVariable Long id, Model model) {
-        customerService.removeCustomerById(id);
+
+        if (!bookingService.getBookingByCustomerId(id)) {
+            customerService.removeCustomerById(id);
+        }
         return handleCustomers(model);
     }
 
     //Skapa senare upp en för ID ist för email vid behov.
     @RequestMapping("/{email}/update")
     public String updateCustomerHandler(@PathVariable String email, Model model){
-        //TODO ska de va en customer? översätter tillbaka till customer
-        Customer c = CustomerConverter.SimpleCustomerDTOtoCustomer(customerService.getCustomerByEmail(email));
+        SimpleCustomerDTO c = customerService.getCustomerByEmailSimpleDTO(email); //Steg 1
+
         model.addAttribute("kund", c);
         return "updateCustomers";
     }
 
     //Temp också
     @PostMapping("/handle/update")
-    public String handleCustomersUpdate(Model model, Customer customer){
+    public String handleCustomersUpdate(Model model, SimpleCustomerDTO customer){
         customerService.updateCustomer(customer);
         int currentPage = 1;
         Page<SimpleCustomerDTO> c = customerService.getAllCustomersPage(currentPage);
@@ -91,7 +104,7 @@ public class CustomerController {
         return "handleCustomers.html";
     }
 
-    //Temp metod nedan. Tas bort sen.
+    //Temp metod nedan. Tas bort sen/justeras.
     @RequestMapping("/frontPage")
         public String loadFrontPageTest(){
             return "Index.html";
